add auto update script

+++ update.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 # set -x
 # set -e
+function user_exists() { id "$1" &>/dev/null; }
+
 function update_pkg() {
 	PKG_NAME=$1
 	NEW_VER=$(echo "$2" | sed 's/^[^.0-9]*//')
@@ -18,30 +20,40 @@ function update_pkg() {
 		pushd $PKG_NAME
 		sed "s/^pkgver=.*$/pkgver=${NEW_VER}/" -i PKGBUILD
 		sed "s/^pkgrel=.*$/pkgrel=1/" -i PKGBUILD
-		updpkgsums
-		makepkg --printsrcinfo > .SRCINFO
+		if user_exists makepkg; then
+			echo "run makepkg as user makepkg"
+			su makepkg -c "updpkgsums"
+			su makepkg -c "makepkg --printsrcinfo > .SRCINFO"
+		else
+			updpkgsums
+			makepkg --printsrcinfo > .SRCINFO
+		fi
 		popd
 		git add $PKG_NAME
 		git commit -m "update $PKG_NAME from $OLD_VER to $NEW_VER"
-		# aurpublish $PKG_NAME
+		aurpublish $PKG_NAME -s
 		;;
 	esac
 }
 
-function check_ver() {
+# use nvchecker to check new versions
+if [[ -f key.toml ]]; then
+	nvchecker --logger both -c nvchecker.toml -k key.toml | tee nvchecker.log
+else
 	nvchecker --logger both -c nvchecker.toml | tee nvchecker.log
-	nvtake -c nvchecker.toml --all
-	git add *.json
-	if git diff --exit-code >/dev/null; then
-		git commit -m "nvchecker update"
-	fi
+fi
 
-}
-
-check_ver
-
+# update pkgbuilds
 grep '"event": "updated"' nvchecker.log | while read line; do
 	PKG_NAME=$(echo $line | jq -r '.name')
 	PKG_VER=$(echo $line | jq -r '.version')
 	update_pkg $PKG_NAME $PKG_VER
 done
+
+# commit nvchecker changes
+nvtake -c nvchecker.toml --all
+git add .
+if git diff --exit-code >/dev/null; then
+	git commit -m "nvchecker update"
+fi
+
